fetch/metadata/generated/script-module-import-static-sec-fetch-site.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/cases/master.yml
- fetch/metadata/tools/templates/script-module-import-static.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTPS</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url) {
    const script = document.createElement('script');
    script.setAttribute('type', 'module');
    script.setAttribute(
      'src',
      '/fetch/metadata/resources/es-module.sub.js?moduleId=' + encodeURIComponent(url)
    );

    return new Promise((resolve, reject) => {
        script.onload = resolve;
        script.onerror = () => reject('Failed to load script');
        document.body.appendChild(script);
      })
      .then(() => script.remove());
  }

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameOrigin', 'crossSite', 'sameOrigin'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Origin -> Cross-Site -> Same-Origin redirect');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['crossSite', 'sameOrigin'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Same Origin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['crossSite', 'sameSite'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Same-Site');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['crossSite', 'crossSite'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Cross-Site');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameOrigin', 'sameOrigin'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'Same-Origin -> Same Origin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameOrigin', 'sameSite'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Origin -> Same-Site');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameOrigin', 'crossSite'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Origin -> Cross-Site');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameSite', 'sameOrigin'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Site -> Same Origin');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameSite', 'sameSite'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Site -> Same-Site');

  promise_test(() => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(
          key, ['sameSite', 'crossSite'], { mime: 'application/javascript' }
        )
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Site -> Cross-Site');
  </script>
</html>
